@@ -12,66 +12,71 @@
   "license": "MIT",
   "repository": "https://github.com/Neufund/Typechain",
   "devDependencies": {
+    "@types/bignumber.js": "^5.0.0",
     "@types/bluebird": "^3.5.18",
     "@types/chai": "^4.0.6",
     "@types/chai-as-promised": "^7.1.0",
-    "@types/chalk": "^0.4.31",
     "@types/command-line-args": "^4.0.2",
     "@types/debug": "^0.0.30",
-    "@types/fs-extra": "^4.0.2",
-    "@types/glob": "^5.0.32",
+    "@types/fs-extra": "^5.0.4",
+    "@types/glob": "^5.0.35",
     "@types/mocha": "^2.2.44",
     "@types/node": "^8.0.25",
-    "@types/prettier": "^1.8.1",
+    "@types/prettier": "^1.13.2",
     "bignumber.js": "^5.0.0",
     "bluebird": "^3.5.1",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "coveralls": "^3.0.2",
-    "ganache-cli": "^6.0.3",
+    "ganache-cli": "^6.1.8",
+    "glob": "^7.1.2",
     "krzkaczor-solc": "^0.4.24",
     "mocha": "^4.0.1",
     "mocha-prepare": "^0.1.0",
     "nyc": "^12.0.2",
+    "patch-package": "^5.1.1",
+    "prettier": "^1.14.2",
+    "snap-shot-it": "^6.1.1",
     "source-map-support": "^0.5.8",
+    "truffle": "^4.1.14",
+    "truffle-typings": "^1.0.1",
     "ts-node": "^3.3.0",
-    "tslint": "^5.8.0",
+    "tslint": "^5.11.0",
     "tslint-language-service": "^0.9.7",
-    "typescript": "^2.6.1",
+    "typescript": "^3.0.1",
     "typestrict": "^0.0.6",
     "web3": "0.20.2",
     "web3-typescript-typings": "^0.7.2"
   },
   "dependencies": {
-    "chalk": "^2.3.0",
     "command-line-args": "^4.0.7",
     "debug": "^3.0.1",
-    "fs-extra": "^4.0.2",
-    "glob": "^7.1.2",
-    "prettier": "^1.9.1"
+    "fs-extra": "^7.0.0",
+    "ts-generator": "^0.0.5"
   },
   "peerDependencies": {
     "bignumber.js": "^5.0.0",
     "web3": "0.20.x"
   },
   "scripts": {
-    "prepublish": "yarn build",
+    "prepare": "patch-package && yarn build",
     "build": "rm -rf ./dist && tsc -p ./tsconfig.production.json",
-    "postbuild": "chmod +x ./dist/cli.js",
+    "postbuild": "chmod +x ./dist/cli/cli.js",
     "lint": "yarn formatting && yarn tslint",
     "lint:fix": "yarn tslint:fix && yarn formatting:fix",
     "tslint": "tslint -c ./tslint.json --project ./tsconfig.json --format stylish",
     "tslint:fix": "tslint -c ./tslint.json --project ./tsconfig.json --format stylish --fix ",
     "tsc": "tsc --noEmit",
+    "tsc:truffle": "tsc --noEmit --project ./test/integration/targets/truffle/tsconfig.json",
     "formatting": "prettier --list-different '**/*.ts'",
     "formatting:fix": "prettier --write '**/*.ts'",
     "test": "yarn lint && ./scripts/test.sh",
     "test:coverage": "yarn lint && ./scripts/test.sh COVERAGE",
     "test:fix": "yarn lint:fix && ./scripts/test.sh",
-    "test:mocha": "NODE_ENV=test mocha --require ts-node/register.js --require test/setup-chai.ts --require 'test/integration/before.ts' 'test/integration/web3.ts' 'test/**/*.spec.ts'",
-    "test:mocha:coverage": "NODE_ENV=test nyc mocha --require ts-node/register.js --require source-map-support/register --require test/setup-chai.ts --require 'test/integration/before.ts' 'test/integration/web3.ts' 'test/**/*.spec.ts'",
+    "test:mocha": "NODE_ENV=test mocha --require ts-node/register.js --require test/setup-chai.ts --require 'test/integration/before.ts' 'test/integration/web3.ts' './**/*.spec.ts'",
+    "test:mocha:coverage": "NODE_ENV=test nyc mocha --require ts-node/register.js --require source-map-support/register --require test/setup-chai.ts --require 'test/integration/before.ts' 'test/integration/web3.ts' './**/*.spec.ts'",
     "coveralls": "nyc report --reporter=text-lcov | coveralls"
   },
-  "main": "./dist/generateTypeChainWrappers.js",
+  "main": "./dist/index.js",
   "bin": "./dist/cli.js"
 }
