@@ -1,24 +1,57 @@
-var prepare = require("mocha-prepare");
+const prepare = require("mocha-prepare");
+import { removeSync } from "fs-extra";
 
-import { generateTypeChainWrappers } from "../../lib/generateTypeChainWrappers";
+import { tsGenerator } from "ts-generator";
 import { join } from "path";
+import { Typechain, ITypechainCfg } from "../../lib";
+import { TPluginCfg } from "ts-generator/dist/parseConfigFile";
+import { readFileSync } from "fs";
 
 /**
  * NOTE: this is done here only to easily count code coverage.
  * Normally you would run typechain in separate build step, before running your tests.
  */
 
-prepare(async (done: any) => {
-  await generateTypeChainWrappers({
-    glob: "**/*.abi",
-    force: true,
-  });
+prepare((done: any) => {
+  (async () => {
+    const cwd = __dirname;
+    const prettierCfg = JSON.parse(readFileSync(join(__dirname, "../../.prettierrc"), "utf8"));
 
-  const outputPath = join(__dirname, "../../test-tmp/");
-  await generateTypeChainWrappers({
-    glob: "**/*.abi",
-    force: true,
-    outDir: outputPath,
+    await generateLegacy(cwd, prettierCfg);
+    await generateTruffle(cwd, prettierCfg);
+
+    done();
+  })().catch(e => {
+    // tslint:disable-next-line
+    console.error(e);
+    process.exit(1);
   });
-  done();
 });
+
+async function generateLegacy(cwd: string, prettierCfg: any) {
+  const outDir = "./targets/legacy/wrappers";
+
+  removeSync(join(__dirname, outDir));
+
+  const rawConfig: TPluginCfg<ITypechainCfg> = {
+    files: "**/*.abi",
+    target: "legacy",
+    outDir,
+  };
+
+  await tsGenerator({ cwd, prettier: prettierCfg }, new Typechain({ cwd, rawConfig }));
+}
+
+async function generateTruffle(cwd: string, prettierCfg: any) {
+  const outDir = "./targets/truffle/@types/truffle-contracts";
+
+  removeSync(join(__dirname, outDir));
+
+  const rawConfig: TPluginCfg<ITypechainCfg> = {
+    files: "targets/truffle/build/**/*.json",
+    target: "truffle",
+    outDir,
+  };
+
+  await tsGenerator({ cwd, prettier: prettierCfg }, new Typechain({ cwd, rawConfig }));
+}
