@@ -1,13 +1,16 @@
 #!/usr/bin/env node
+import { tsGenerator } from "ts-generator";
+
 import { parseArgs } from "./parseArgs";
-import { generateTypeChainWrappers } from "./generateTypeChainWrappers";
-import { logger } from "./logger";
+import { Typechain } from "..";
+import { logger } from "../utils/logger";
 
 async function main() {
   (global as any).IS_CLI = true;
   const options = parseArgs();
+  const cwd = process.cwd();
 
-  await generateTypeChainWrappers(options);
+  await tsGenerator({ cwd }, new Typechain({ cwd, rawConfig: options }));
 }
 
 main().catch(e => {
