@@ -1,6 +1,5 @@
 import {
   RawAbiDefinition,
-  parse,
   Contract,
   AbiParameter,
   EventArgDeclaration,
@@ -8,61 +7,70 @@ import {
   ConstantFunctionDeclaration,
   FunctionDeclaration,
   EventDeclaration,
-} from "./abiParser";
-import { getVersion } from "./utils";
-import { EvmType, ArrayType } from "./typeParser";
+  parse,
+} from "../../parser/abiParser";
+import {
+  EvmType,
+  ArrayType,
+  BooleanType,
+  IntegerType,
+  UnsignedIntegerType,
+  VoidType,
+  StringType,
+  BytesType,
+  AddressType,
+} from "../../parser/typeParser";
+import { IContext } from "../shared";
+import { join } from "path";
+import { readFileSync } from "fs";
 
-export interface IContext {
-  fileName: string;
-  relativeRuntimePath: string;
+export function getRuntime(): string {
+  const runtimePath = join(__dirname, "./runtime/typechain-runtime.ts");
+  return readFileSync(runtimePath, "utf8");
 }
 
-export function generateSource(abi: Array<RawAbiDefinition>, context: IContext): string {
-  const parsedContractAbi = parse(abi);
+export function codegen(abi: Array<RawAbiDefinition>, context: IContext): string {
+  const parsedContractAbi = parse(abi, context.fileName);
 
   return codeGenForContract(abi, parsedContractAbi, context);
 }
 
-// @todo better typings for web3
 function codeGenForContract(abi: Array<RawAbiDefinition>, input: Contract, context: IContext) {
   const runtimeNamespace = "TC";
   const typeName = context.fileName;
 
   return `
-/* GENERATED BY TYPECHAIN VER. ${getVersion()} */
-/* tslint:disable */
-  
-import { BigNumber } from "bignumber.js";
-import * as ${runtimeNamespace} from "${context.relativeRuntimePath}"
+    import { BigNumber } from "bignumber.js";
+    import * as ${runtimeNamespace} from "${context.relativeRuntimePath}"
 
-export class ${typeName} extends ${runtimeNamespace}.TypeChainContract {
-    public readonly rawWeb3Contract: any;
-  
-    public constructor(web3: any, address: string | BigNumber) {
-      const abi = ${JSON.stringify(abi)};
-      super(web3, address, abi);
-    }
+    export class ${typeName} extends ${runtimeNamespace}.TypeChainContract {
+      public readonly rawWeb3Contract: any;
 
-    static async createAndValidate(web3: any, address: string | BigNumber): Promise<${typeName}> {
-      const contract = new ${typeName}(web3, address);
-      const code = await ${runtimeNamespace}.promisify(web3.eth.getCode, [address]);
-
-      // in case of missing smartcontract, code can be equal to "0x0" or "0x" depending on exact web3 implementation
-      // to cover all these cases we just check against the source code length — there won't be any meaningful EVM program in less then 3 chars
-      if (code.length < 4) {
-        throw new Error(\`Contract at \${address} doesn't exist!\`);
+      public constructor(web3: any, address: string | BigNumber) {
+        const abi = ${JSON.stringify(abi)};
+        super(web3, address, abi);
       }
-      return contract; 
+
+      static async createAndValidate(web3: any, address: string | BigNumber): Promise<${typeName}> {
+        const contract = new ${typeName}(web3, address);
+        const code = await ${runtimeNamespace}.promisify(web3.eth.getCode, [address]);
+
+        // in case of missing smartcontract, code can be equal to "0x0" or "0x" depending on exact web3 implementation
+        // to cover all these cases we just check against the source code length — there won't be any meaningful EVM program in less then 3 chars
+        if (code.length < 4) {
+          throw new Error(\`Contract at \${address} doesn't exist!\`);
+        }
+        return contract; 
+      }
+      
+      ${codeGenForConstants(runtimeNamespace, input.constants)}
+      
+      ${codeGenForConstantsFunctions(runtimeNamespace, input.constantFunctions)}
+      
+      ${codeGenForFunctions(runtimeNamespace, input.functions)}
+      
+      ${codeGenForEvents(runtimeNamespace, input.events)}
     }
-    
-    ${codeGenForConstants(runtimeNamespace, input.constants)}
-    
-    ${codeGenForConstantsFunctions(runtimeNamespace, input.constantFunctions)}
-    
-    ${codeGenForFunctions(runtimeNamespace, input.functions)}
-    
-    ${codeGenForEvents(runtimeNamespace, input.events)}
-  }
 `;
 }
 
@@ -70,7 +78,7 @@ function codeGenForConstants(runtimeNamespace: string, constants: Array<Constant
   return constants
     .map(
       ({ name, output }) => `
-        public get ${name}(): Promise<${output.generateCodeForOutput()}> { 
+        public get ${name}(): Promise<${codeGenForOutput(output)}> { 
             return ${runtimeNamespace}.promisify(this.rawWeb3Contract.${name}, []); 
         }
     `,
@@ -133,7 +141,7 @@ function codeGenForEvents(runtimeNamespace: string, events: Array<EventDeclarati
 }
 
 function codeGenForParams(param: AbiParameter, index: number): string {
-  return `${param.name || `arg${index}`}: ${param.type.generateCodeForInput()}`;
+  return `${param.name || `arg${index}`}: ${codeGenForInput(param.type)}`;
 }
 
 function codeGenForArgs(param: AbiParameter, index: number): string {
@@ -144,9 +152,9 @@ function codeGenForArgs(param: AbiParameter, index: number): string {
 
 function codeGenForOutputTypeList(output: Array<EvmType>): string {
   if (output.length === 1) {
-    return output[0].generateCodeForOutput();
+    return codeGenForOutput(output[0]);
   } else {
-    return `[${output.map(x => x.generateCodeForOutput()).join(", ")}]`;
+    return `[${output.map(x => codeGenForOutput(x)).join(", ")}]`;
   }
 }
 
@@ -154,11 +162,49 @@ function codeGenForEventArgs(args: EventArgDeclaration[], onlyIndexed: boolean) 
   return `{${args
     .filter(arg => arg.isIndexed || !onlyIndexed)
     .map(arg => {
-      const inputCodegen = arg.type.generateCodeForInput();
+      const inputCodegen = codeGenForInput(arg.type);
 
       // if we're specifying a filter, you can take a single value or an array of values to check for
       const argType = `${inputCodegen}${onlyIndexed ? ` | Array<${inputCodegen}>` : ""}`;
       return `${arg.name}${onlyIndexed ? "?" : ""}: ${argType}`;
     })
     .join(`, `)}}`;
+}
+
+function codeGenForInput(evmType: EvmType): string {
+  switch (evmType.constructor) {
+    case IntegerType:
+      return "BigNumber | number";
+    case UnsignedIntegerType:
+      return "BigNumber | number";
+    case AddressType:
+      return "BigNumber | string";
+
+    default:
+      return codeGenForOutput(evmType);
+  }
+}
+
+function codeGenForOutput(evmType: EvmType): string {
+  switch (evmType.constructor) {
+    case BooleanType:
+      return "boolean";
+    case IntegerType:
+      return "BigNumber";
+    case UnsignedIntegerType:
+      return "BigNumber";
+    case VoidType:
+      return "void";
+    case StringType:
+      return "string";
+    case BytesType:
+      return "string";
+    case AddressType:
+      return "string";
+    case ArrayType:
+      return codeGenForOutput((evmType as ArrayType).itemType) + "[]";
+
+    default:
+      throw new Error(`Unrecognized ABI piece: ${evmType.constructor}`);
+  }
 }
