@@ -1,8 +1,10 @@
 import { expect } from "chai";
-import { deployContract } from "./utils/web3Contracts";
+import { deployContract } from "../../utils/web3Contracts";
 
-import { web3 } from "./web3";
-import { ContractWithOverloads } from "./abis/ContractWithOverloads";
+import { web3 } from "../../web3";
+import { ContractWithOverloads } from "./wrappers/ContractWithOverloads";
+import { snapshotSource } from "../../utils/snapshotSource";
+import { join } from "path";
 
 describe("ContractWithOverloads", () => {
   let contractAddress: string;
@@ -10,6 +12,9 @@ describe("ContractWithOverloads", () => {
   beforeEach(async () => {
     contractAddress = (await deployContract("ContractWithOverloads")).address;
   });
+
+  it("should snapshot generated code", () =>
+    snapshotSource(join(__dirname, "./wrappers/ContractWithOverloads.ts")));
 
   it("should be able to access counter", async () => {
     const contractWithOverloads = await ContractWithOverloads.createAndValidate(
